refactor charts panel header to own component
